refactor(server/app.js): seperated graphql part
