+++ server/app.js
@@ -2,8 +2,7 @@ const express = require("express");
 const path = require("path");
 const cookieParser = require("cookie-parser");
 const logger = require("morgan");
-const { gql, ApolloServer } = require("apollo-server-express");
-const axios = require("axios");
+const { ApolloServer } = require("apollo-server-express");
 const { generateToken, authorize } = require("./util/jwt");
 const indexRouter = require("./routes/index");
 const usersRouter = require("./routes/users");
@@ -21,112 +20,8 @@ app.use("/", indexRouter);
 app.use("/users", usersRouter);
 app.use("/auth", authRouter);
 
-//graphql
-const typeDefs = gql`
-  type Owner {
-    id: ID!
-    login: String!
-    url: String
-    avatarUrl: String
-  }
-
-  type Nodes {
-    name: String!
-    owner: Owner!
-    stargazerCount: Int
-  }
-
-  type RepositoriesContributedTo {
-    nodes: [Nodes!]
-  }
-
-  type Followers {
-    totalCount: Int
-  }
-
-  type Following {
-    totalCount: Int
-  }
-
-  type User {
-    login: String!
-    name: String!
-    repositoriesContributedTo: RepositoriesContributedTo
-    bio: String
-    email: String
-    followers: Followers
-    following: Following
-    avatarUrl: String
-    id: ID!
-  }
-
-  type Query {
-    user(login: String): User
-  }
-`;
-
-const resolvers = {
-  Query: {
-    user: async (root, args, context) => {
-      var data = JSON.stringify({
-        query: `query {
-          user(login:"${args.login}") {
-            login
-            name
-            repositoriesContributedTo(contributionTypes: PULL_REQUEST, first: 50, orderBy: {field: STARGAZERS, direction: DESC}) {
-              nodes {
-                name
-                owner {
-                  id
-                  login
-                  url
-                  avatarUrl
-                }
-                stargazerCount
-              }
-            }
-            bio
-            email
-            followers {
-              totalCount
-            }
-            following {
-              totalCount
-            }
-            avatarUrl
-            id
-          }
-        }
-        `,
-        variables: {},
-      });
-
-      var config = {
-        method: "post",
-        url: "https://api.github.com/graphql",
-        headers: {
-          Authorization: `Bearer ${context.access_token}`,
-          "Content-Type": "application/json",
-        },
-        data: data,
-      };
-
-      let name = "not found";
-
-      await axios(config)
-        .then((response) => {
-          name = response.data.data.user;
-          console.log(args.login);
-        })
-        .catch((error) => {
-          console.log(error);
-          console.log(response.data);
-        });
-
-      return name;
-    },
-  },
-};
+const typeDefs = require('./graphql/typeDefs');
+const resolvers = require('./graphql/resolver');
 
 app.use(authorize);
 
@@ -137,6 +32,7 @@ const server = new ApolloServer({
     access_token: req.access_token || "",
   }),
 });
+
 server.applyMiddleware({ app });
 
 module.exports = app;
